Add missing selenium driver information
