+++ src/test/java/e2e/pages/LoginPage.java
@@ -1,16 +1,15 @@
-package e2e;
+package e2e.pages;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 
-public class RegisterPage extends BasePage {
+public class LoginPage extends BasePage {
 
-    public RegisterPage() {
+    public LoginPage() throws Exception {
         super();
-        path = "register";
+        path = "login";
+        uniqueId = "loginPageUnique";
     }
-
-
 
     public WebElement getUsernameField() {
         return driver.findElement(By.id("username"));
@@ -20,8 +19,8 @@ public class RegisterPage extends BasePage {
         return driver.findElement(By.id("password"));
     }
 
-    public WebElement getRegisterButton() {
-        return driver.findElement(By.id("registerButton"));
+    public WebElement getLoginButton() {
+        return driver.findElement(By.id("loginButton"));
     }
 
     public WebElement getErrorAlert() {
