+++ src/test/java/e2e/pages/MainPage.java
@@ -1,8 +1,9 @@
-package e2e;
+package e2e.pages;
 
 public class MainPage extends BasePage {
 
-    public MainPage() {
+    public MainPage() throws Exception {
+        super();
         uniqueId = "mainPageUnique";
     }
 }
